third_party/WebKit/LayoutTests/http/tests/inspector/network/network-preview-json.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>

function test()
{
    function check(text) {
        InspectorTest.addResult("");
        InspectorTest.addResult("Input: " + text);
        var parsedJSON = WebInspector.RequestJSONView.parseJSON(text);
        if (!parsedJSON) {
            InspectorTest.addResult("Can't parse");
            return;
        }
        InspectorTest.addResult("Prefix: " + parsedJSON.prefix);
        InspectorTest.addResult("Data: " + JSON.stringify(parsedJSON.data));
        InspectorTest.addResult("Suffix: " + parsedJSON.suffix);
    }

    check("{\"name\": \"value\"}");
    check("while(1); {\"name\": \"value\"}");
    check("[,\"foo\", -4.2, true, false, null]");
    check("[{\"foo\": {}, \"bar\": []},[[],{}]]");
    check("/* vanilla */ run([1, 2, 3]);");
    check("[\"A\\\"B\\u0020C\\nD\\\\E\\u04ABF\"]");

    check("<html>404 Page not found</html>");

    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests RequestJSONView ability to parse JSON passed in XHR, JSONP</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=65559">Bug 65559</a>
</body>
</html>
